+++ CRM/DataprocessorOutputExport/CSV.php
@@ -195,19 +195,19 @@ class CRM_DataprocessorOutputExport_CSV implements ExportOutputInterface, Direct
 
 
   protected function startBatchJob(\Civi\DataProcessor\ProcessorType\AbstractProcessorType $dataProcessorClass, $dataProcessor, $outputBAO, $formValues, $sortFieldName = null, $sortDirection = 'ASC', $idField=null, $selectedIds=array()) {
-    $session = CRM_Core_Session::singleton();
+    $session = \CRM_Core_Session::singleton();
 
-    $name = date('Ymdhis').'_'.$dataProcessor['id'].'_'.$outputBAO['id'].'_'.CRM_Core_Session::getLoggedInContactID().'_'.$dataProcessor['name'];
+    $name = date('Ymdhis').'_'.$dataProcessor['id'].'_'.$outputBAO['id'].'_'.CRM_Core_Session::getLoggedInContactID().'_'.md5($dataProcessor['name']);
 
-    $queue = CRM_Queue_Service::singleton()->create(array(
+    $queue = \CRM_Queue_Service::singleton()->create(array(
       'type' => 'Sql',
       'name' => $name,
       'reset' => TRUE, //do flush queue upon creation
     ));
 
-    $basePath = CRM_Core_Config::singleton()->templateCompileDir . 'dataprocessor_export_csv';
-    CRM_Utils_File::createDir($basePath);
-    CRM_Utils_File::restrictAccess($basePath.'/');
+    $basePath = \CRM_Core_Config::singleton()->templateCompileDir . 'dataprocessor_export_csv';
+    \CRM_Utils_File::createDir($basePath);
+    \CRM_Utils_File::restrictAccess($basePath.'/');
     $filename = $basePath.'/'. $name.'.csv';
 
     self::createHeaderLine($filename, $dataProcessorClass, $outputBAO['configuration']);
@@ -221,7 +221,7 @@ class CRM_DataprocessorOutputExport_CSV implements ExportOutputInterface, Direct
       ));
 
       //create a task without parameters
-      $task = new CRM_Queue_Task(
+      $task = new \CRM_Queue_Task(
         array(
           'CRM_DataprocessorOutputExport_CSV',
           'exportBatch'
@@ -235,7 +235,7 @@ class CRM_DataprocessorOutputExport_CSV implements ExportOutputInterface, Direct
 
     $url = str_replace("&amp;", "&", $session->readUserContext());
 
-    $runner = new CRM_Queue_Runner(array(
+    $runner = new \CRM_Queue_Runner(array(
       'title' => E::ts('Exporting data'), //title fo the queue
       'queue' => $queue, //the queue object
       'errorMode'=> CRM_Queue_Runner::ERROR_CONTINUE, //abort upon error and keep task in queue