+++ Civi/DataProcessor/DataFlow/AbstractDataFlow.php
@@ -80,6 +80,7 @@ abstract class AbstractDataFlow {
    */
   protected function resetInitializeState() {
     $this->currentRecordIndex = 0;
+    unset($this->_allRecords);
   }
 
   /**
@@ -147,7 +148,7 @@ abstract class AbstractDataFlow {
    * @return array
    */
   public function allRecords($fieldNameprefix = '') {
-    if (!is_array($this->_allRecords)) {
+    if (!isset($this->_allRecords) || !is_array($this->_allRecords)) {
       $this->_allRecords = [];
       $_allRecords = [];
       try {